Share 'ProductItem' class with all scrapers
