+++ src/scraper_biocoop/products_spider.py
@@ -5,8 +5,7 @@ from scrapy import Request, Spider
 
 from models.category import CategoryValues
 from models.product import QuantityUnit
-
-from .items import ProductItem
+from utils.spider import ProductItem
 
 
 class BiocoopProductsSpider(Spider):
